+++ R/dictionary_tidiers.R
@@ -10,8 +10,8 @@
 #'
 #' @export
 tidy.dictionary <- function(x, regex = FALSE, ...) {
-  ret <- plyr::ldply(x, function(e) data.frame(word = e, stringsAsFactors = FALSE),
-                     .id = "category") %>%
+  ret <- purrr::map_df(x, function(e) data.frame(word = e, stringsAsFactors = FALSE),
+                       .id = "category") %>%
     mutate(category = as.character(category))
 
   if (regex) {