+++ NEWS.md
@@ -5,6 +5,7 @@
 * Add tidiers for LDA objects from topicmodels
 * Fixed a bug when tidying by line/sentence/paragraph/regex and there are multiple non-text columns
 * Added ability to pass a (custom tokenizing) function to token. Also added a collapse argument that makes the choice whether to combine lines before tokenizing explicit.
+* Changed tidy.dictionary to return a tbl_df rather than a data.frame
 
 # tidytext 0.1.0